+++ mobile/lib/data/models/stock_model.dart
@@ -2,7 +2,7 @@ class StockModel {
   final int id;
   final int projectId;
   final int materialId;
-  final double availableQuantity;
+  final int availableQuantity;
   final String? materialName;
   final String? materialUnit;
   final String? projectName;
@@ -30,8 +30,8 @@ class StockModel {
       projectId: (json['project_id'] as num?)?.toInt() ?? 0,
       materialId: (json['material_id'] as num?)?.toInt() ?? 0,
       availableQuantity: (json['available_quantity'] is String)
-          ? double.tryParse(json['available_quantity']) ?? 0.0
-          : (json['available_quantity'] as num?)?.toDouble() ?? 0.0,
+          ? int.tryParse(json['available_quantity']) ?? 0
+          : (json['available_quantity'] as num?)?.toInt() ?? 0,
       materialName: material?['name'] as String?,
       materialUnit: material?['unit'] as String?,
       projectName: project?['name'] as String?,
